Moved more things from core to model
