+++ src/model/commands.scala
@@ -15,7 +15,7 @@
    ╚═══════════════════════════════════════════════════════════════════════════════════════════════════════════╝
 */
 
-package fury.core
+package fury.model
 
 import fury.io._, fury.strings._
