+++ src/model/package.scala
@@ -19,6 +19,9 @@ package fury
 
 import fury.strings._, fury.io._
 
+import gastronomy._
+
 package object model {
-  
+  implicit val fileSystemSafeBase64Url: ByteEncoder[Base64Url] =
+    ByteEncoder.base64.encode(_).replace('/', '_').takeWhile(_ != '=')
 }
